finally done with making_the_grade.cpp excercise

+++ Excercism/freelancer_rates.cpp
@@ -22,8 +22,7 @@ double apply_discount(double before_discount, double discount)
 int monthly_rate(double hourly_rate, double discount)
 {
     // TODO: Implement a function to calculate the monthly rate, and apply a
-    // discount.
-    return std::ceil((daily_rate(apply_discount(hourly_rate, discount))) * 22.0);
+    return static_cast<int>(std::ceil(daily_rate(apply_discount(hourly_rate, discount)) * 22.0));
 }
 
 // days_in_budget calculates the number of workdays given a budget, hourly rate,
@@ -33,5 +32,5 @@ int days_in_budget(int budget, double hourly_rate, double discount)
 {
     // TODO: Implement a function that takes a budget, an hourly rate, and a
     // discount, and calculates how many complete days of work that covers.
-    return budget / daily_rate(apply_discount(hourly_rate, discount));
+    return static_cast<int>(budget / daily_rate(apply_discount(hourly_rate, discount)));
 }
